110 - Refactor edit profile page, and change navigation
