+++ app/src/pages/profile/EditProfileFields/PhoneFIeld/styles.js
@@ -1,7 +1,6 @@
 import { StyleSheet } from 'react-native';
-import colors from '../../../../assets/styles/colorVariables';
-import fonts from '../../../../assets/styles/fontVariable';
-
+import colors from '../../../../../assets/styles/colorVariables';
+import fonts from '../../../../../assets/styles/fontVariable';
 export default StyleSheet.create({
     container: {
         flex: 1,
